requirements.txt
databricks-sql-connector>=2.9.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=23.1.0
cachetools>=5.3.0
click>=8.1.0
openai>=1.50.0
requests>=2.32.0

